Velocidad mejorada

+++ Game.hpp
@@ -17,7 +17,7 @@
 class Game {
     float speed;
     float acceleration;
-    int posX;
+    float posX;
 
     std::vector<std::vector<Map>> maps; // 5 blocks of 15 maps {(0), (1, 2), (3, 4, 5), (6, 7, 8, 9), (10, 11, 12, 13, 14))}
     std::pair<int, int> mapId; // (Block, num. map), ex: map0 = (0, 0); map1 = (1, 0); map2 = (1, 1); map14 = (4, 4)
